Refactor: 창고 크기 타입 변경

+++ src/main/java/com/acorn/api/entity/container/Container.java
@@ -32,7 +32,7 @@ public class Container {
 
     private String containerContentsText;
 
-    private Integer containerSize;
+    private BigDecimal containerSize;
 
     private Integer containerPrice;
